requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
Deprecated==1.2.14
Django==4.2.5
fonttools==4.42.1
idna==3.4
importlib-resources==6.0.1
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.7.2
MLB-StatsAPI==1.6.2
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pyarrow==13.0.0
pybaseball @ git+https://github.com/jldbc/pybaseball.git@06fa0daa20e1231f15a187337a352377f2c1f766
pycparser==2.21
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wrapt==1.15.0
zipp==3.16.2
